Create 2D arbitrary density sampler
Sample from any arbitrary density function in 2D. The toy model for the energy function is a squared 2d fourier cosine series. The target density is the exponential of the negative squared 2d fcs multiplied by a temperature factor. 

In the end, the normalized analytic density used in the log likelihood function is compared with the kernel density estimate of all the returned samples using a rectangular uniform prior and the energy function in the log like function as exp(-V)

The partition function Z is also used to display P* = exp(-V) / Z